requirements.txt
SQLAlchemy~=2.0.23
python-dotenv~=1.0.0
fastapi~=0.104.1
pydantic~=2.4.2
uvicorn~=0.23.2
python-jose~=3.3.0
passlib~=1.7.4
bcrypt~=4.0.1